requirement.txt
asgiref==3.8.1
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
pip==25.1.1
pyjwt==2.9.0
python-dotenv==1.1.1
setuptools==65.5.0
sqlparse==0.5.3
typing-extensions==4.14.0
whitenoise==6.9.0
